refactor / fix stuff
