+++ modal-js/src/index.ts
@@ -5,8 +5,8 @@ export {
   type LookupOptions,
   type SandboxCreateOptions,
 } from "./app";
+export { type ClientOptions, initializeClient } from "./client";
 export { Cls, ClsInstance } from "./cls";
-export { type ClientOptions, initializeClient } from "./config";
 export {
   FunctionTimeoutError,
   RemoteError,
